+++ src/utils/utils.ts
@@ -1,9 +1,10 @@
 
 export const isEmpty = (value: unknown) =>
-  value === undefined ||
-  value === null ||
-  (typeof value === 'object' && Object.keys(value).length === 0) ||
-  (typeof value === 'string' && value.trim().length === 0);
+  value === undefined
+  || value === null
+  || value === false
+  || (typeof value === 'object' && Object.keys(value).length === 0)
+  || (typeof value === 'string' && value.trim().length === 0);
 
 
 export const snakeToCamelCase = (str: string) => str.replace(